Fix tests
Aside from updating tests for the new iso3 functions, the function mockers had a bug.
Specific error: running test_rollout_grids.py by itself is ok, but running the tests as a whole will fail.
Root cause was found to have been the mock patches for function return values in test_geoboundaries and test_iso3.
It was hard to find a fix, so decided to remove the mocking instead.
The functions run quickly anyway, and this way, we can actually test real API return values.